Genesis Commit
Great code for the CSS class switch and HTML page settings saver to the LocalMemory.

Can call the ClassSwitch(TagName, ClassName, MemoryName) with parameters, that will create custom-named MemoryName memory setting variable in the localMemory automatically. Could be a security risk, because it can also overwrite existing memory cells. So, do not save sensitive things in LocalMemory :-)

The CSS class has to exist in the CSS portion of the page or app ! Otherwise it's pointless. There will be no errors, but it will also only work with previously defined CSS classes in the CSS portion of the HTML.