reverse logic for the exception case
In the exception case where: the user sets dark mode by default, but wants to switch to light color.

In this case the html.dark becomes white :-)
It's  like a parallel universe, where the use contradicts itself and the result is the opposite of our logic.

It works out, but our brains are not made to process for how that works. We need to think two or three times to process that logic.