improve switch logic quality against exceptions
The new logic is more universal and more stable.

Now we get in the memory what we see on the display.

+ Fixes issues, when CSS class was set by default and the switch has to change it without making a memory.

+ We now only remember the  CSS class that was given to us by the function call.

+ Fixes memory conflicts, when the memory cell had wrong data, or older switch data.

+ Now we deep clean the switch memory cell on every CSS class switch change, so that no old or wrong data can exist in our switch memory.

This resolves issues, where the memory was manipulated, or had wrong data, or corrupt data, or with manual data input without permission. So, it's allays a fresh switch memory, when we make the new memory. With this we reduce memory errors on the switch side.


// For testing the variables and logic.

 console.log("New Class was : " + ClassName ) ;

 console.log("ClassWasActive was : " + ClassWasActive ) ;

 console.log("OldMemory was : " + OldMemory ) ;


// Have fun future traveler !